# openapi.yaml
openapi: "3.0.3"
info:
  title: flickies
  version: "0.3.0"
  description: |
    Self-hosted video REST + MCP API. Lipsync (Wav2Lip + LatentSync 1.5),
    face restoration (GFPGAN), ffmpeg-driven video ops, async jobs, webhooks.

    The MCP endpoint at /v1/mcp is JSON-RPC over a streamable HTTP
    transport and is NOT described in this document — use an MCP client.

    ## License posture

    Wav2Lip variants are trained on LRS2 (non-commercial). The server refuses
    to load them unless `FLICKIES_ENABLE_NONCOMMERCIAL=1`. LatentSync 1.5
    (Apache-2.0) is the commercial-safe default.

    ## API contract

    Every video endpoint takes a JSON body. Multipart upload is supported
    only on `PUT /v1/files/{path}` — the local-staging interface.

    **Input** — exactly one of:
      - `file_path` (FILES_DIR-relative, server reads from local /v1/files
        staging area)
      - `file_url` (any HTTP/HTTPS URL the server fetches)

    **Output** — exactly one of:
      - `output_path` (server writes to FILES_DIR/<path>, response JSON is
        `{path, size, ...}`; caller downloads via `GET /v1/files/<path>`)
      - `output_url` (server PUTs to the URL — presigned PUT or
        PUT-accepting endpoint; response JSON is `{url, size, ...}`)

    `async_job=true` auto-stages to `jobs/{id}.{ext}` under FILES_DIR if
    neither output is given.
servers:
  - url: http://localhost:8000
security:
  - BearerAuth: []
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
  parameters:
    EngineSlug:
      in: path
      name: slug
      required: true
      schema:
        type: string
      description: Engine slug. Engines are ML-only (`wav2lip`, `wav2lip-gan`, `latentsync-1.5`, `gfpgan`); ffmpeg is the standard CPU backend and isn't an engine.
    FilesPath:
      in: path
      name: path
      required: true
      schema:
        type: string
      description: FILES_DIR-relative path (slashes allowed).
...
  responses:
...
    "401":
      description: Missing or invalid bearer token.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorBody" }
...
    "404":
      description: Engine or file not found.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorBody" }
...
  schemas:
...
    EngineList:
      type: object
      required: [engines]
      properties:
        engines:
          type: array
          items: { $ref: "#/components/schemas/EngineInfo" }

    # ── jobs / files / async ───────────────────────────────────────────────
...
    StagedOutputResponse:
      type: object
      required: [path, size]
      properties:
        path: { type: string }
        size: { type: integer }
        sha256: { type: string, nullable: true }
        duration_sec: { type: number, nullable: true }
        width: { type: integer, nullable: true }
        height: { type: integer, nullable: true }
...
paths:
...
  /v1/engines:
    get:
      tags: [Engines]
      operationId: listEngines
      summary: List registered engines + load state.
      responses:
        "200":
          description: Engine list.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/EngineList" }
        "401": { $ref: "#/components/responses/401" }
  /v1/engines/{slug}:
    delete:
      tags: [Engines]
      operationId: evictEngine
      summary: Evict the engine from VRAM (no-op if not loaded).
      parameters:
        - $ref: "#/components/parameters/EngineSlug"
      responses:
        "204":
          description: Evicted.
        "401": { $ref: "#/components/responses/401" }
        "404": { $ref: "#/components/responses/404" }
  /v1/files/{path}:
    put:
      tags: [Files]
      operationId: putFile
      summary: Upload a video/audio file to the local staging area.
      parameters:
        - $ref: "#/components/parameters/FilesPath"
      requestBody:
        required: true
        content:
          application/octet-stream:
            schema: { type: string, format: binary }
      responses:
        "201":
          description: Uploaded.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/StagedOutputResponse" }
        "401": { $ref: "#/components/responses/401" }
        "413": { $ref: "#/components/responses/413" }
    get:
      tags: [Files]
      operationId: getFile
      summary: Download a staged file.
      parameters:
        - $ref: "#/components/parameters/FilesPath"
      responses:
        "200":
          description: File stream.
          content:
            application/octet-stream:
              schema: { type: string, format: binary }
        "401": { $ref: "#/components/responses/401" }
        "404": { $ref: "#/components/responses/404" }
    delete:
      tags: [Files]
      operationId: deleteFile
      parameters:
        - $ref: "#/components/parameters/FilesPath"
      responses:
        "204":
          description: Deleted.
        "401": { $ref: "#/components/responses/401" }
        "404": { $ref: "#/components/responses/404" }
  /v1/jobs/{job_id}:
    get:
      tags: [Jobs]
      operationId: getJob
      parameters:
        - $ref: "#/components/parameters/JobIdPath"
      responses:
        "200":
          description: Job status.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/JobStatusResponse" }
        "401": { $ref: "#/components/responses/401" }
        "404": { $ref: "#/components/responses/404" }
  /v1/video/lipsync:
    post:
      tags: [Lipsync]
      operationId: postVideoLipsync
      summary: Drive a face from an audio track.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/VideoLipsyncRequest" }
      responses:
        "200":
          description: Sync result.
          content:
            application/json:
              schema:
                oneOf:
                  - $ref: "#/components/schemas/StagedOutputResponse"
                  - $ref: "#/components/schemas/UrlOutputResponse"
        "202":
          description: Async accepted.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/JobAcceptedResponse" }
        "400": { $ref: "#/components/responses/400" }
        "401": { $ref: "#/components/responses/401" }
        "403": { $ref: "#/components/responses/403" }
        "422": { $ref: "#/components/responses/422" }
  /v1/video/restore:
    post:
      tags: [Restore]
      operationId: postVideoRestore
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/VideoRestoreRequest" }
      responses:
        "200":
          description: Sync result.
          content:
            application/json:
              schema:
                oneOf:
                  - $ref: "#/components/schemas/StagedOutputResponse"
                  - $ref: "#/components/schemas/UrlOutputResponse"
        "202":
          description: Async accepted.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/JobAcceptedResponse" }
        "400": { $ref: "#/components/responses/400" }
        "401": { $ref: "#/components/responses/401" }
        "422": { $ref: "#/components/responses/422" }
  /v1/video/trim:
    post:
      tags: [Video Transform]
      operationId: postVideoTrim
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/VideoTrimRequest" }
      responses:
        "200":
          description: Trimmed.
          content:
            application/json:
              schema:
                oneOf:
                  - $ref: "#/components/schemas/StagedOutputResponse"
                  - $ref: "#/components/schemas/UrlOutputResponse"
        "400": { $ref: "#/components/responses/400" }
        "422": { $ref: "#/components/responses/422" }
  /v1/video/concat:
    post:
      tags: [Video Transform]
      operationId: postVideoConcat
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/VideoConcatRequest" }
      responses:
        "200":
          description: Concatenated.
          content:
            application/json:
              schema:
                oneOf:
                  - $ref: "#/components/schemas/StagedOutputResponse"
                  - $ref: "#/components/schemas/UrlOutputResponse"
  /v1/video/transcode:
    post:
      tags: [Video Transform]
      operationId: postVideoTranscode
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/VideoTranscodeRequest" }
      responses:
        "200":
          description: Transcoded.
          content:
            application/json:
              schema:
                oneOf:
                  - $ref: "#/components/schemas/StagedOutputResponse"
                  - $ref: "#/components/schemas/UrlOutputResponse"
  /v1/video/scale:
    post:
      tags: [Video Transform]
      operationId: postVideoScale
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/VideoScaleRequest" }
      responses:
        "200":
          description: Scaled.
          content:
            application/json:
              schema:
                oneOf:
                  - $ref: "#/components/schemas/StagedOutputResponse"
                  - $ref: "#/components/schemas/UrlOutputResponse"
  /v1/video/mux_audio:
    post:
      tags: [Video Transform]
      operationId: postVideoMuxAudio
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/VideoMuxAudioRequest" }
      responses:
        "200":
          description: Muxed.
          content:
            application/json:
              schema:
                oneOf:
                  - $ref: "#/components/schemas/StagedOutputResponse"
                  - $ref: "#/components/schemas/UrlOutputResponse"
  /v1/video/extract_audio:
    post:
      tags: [Video Transform]
      operationId: postVideoExtractAudio
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/VideoExtractAudioRequest" }
      responses:
        "200":
          description: Audio extracted.
          content:
            application/json:
              schema:
                oneOf:
                  - $ref: "#/components/schemas/StagedOutputResponse"
                  - $ref: "#/components/schemas/UrlOutputResponse"
  /v1/video/thumbnail_grid:
    post:
      tags: [Video Transform]
      operationId: postVideoThumbnailGrid
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/VideoThumbnailGridRequest" }
      responses:
        "200":
          description: PNG grid produced.
          content:
            application/json:
              schema:
                oneOf:
                  - $ref: "#/components/schemas/StagedOutputResponse"
                  - $ref: "#/components/schemas/UrlOutputResponse"
  /v1/video/info:
    post:
      tags: [Info]
      operationId: postVideoInfo
      summary: All metadata for a video — duration, codecs, fps, dimensions, bitrate.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/VideoInfoRequest" }
      responses:
        "200":
          description: ffprobe metadata.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/VideoInfoResponse" }
        "400": { $ref: "#/components/responses/400" }
        "401": { $ref: "#/components/responses/401" }